Task4 result;
