+++ src/main/java/home/work/four/Task4.java
@@ -1,15 +1,20 @@
 package home.work.four;
 
 public class Task4 {
-
-    public static int progression (int initial, int step, int quantity){
-        int result = initial;
-        int currentNumber = initial;
-        for (int i = 1; i < quantity; i++) {
-            currentNumber += step;
-            result *= currentNumber;
+    public static int SumGeometricElements(int init, double step, int alim) {
+        int result = 0;
+        int currentValue = init;
+        if (step == 0 || alim < 0 || step < 0) {
+            throw new IllegalArgumentException();
+        } else  {
+            while (currentValue >= alim){
+                result += currentValue;
+                currentValue *= step;
+            }
         }
-        return result;
-    }
 
+
+        return result;
+
+    }
 }
