+++ src/main/java/org/example/Main.java
@@ -1,18 +1,13 @@
 package org.example;
 
-import home.work.four.SortOrder;
-import home.work.four.Task1;
-import home.work.four.Task2;
+import home.work.four.Task3;
 import home.work.four.Task4;
-
-import java.lang.reflect.Array;
-import java.sql.SQLOutput;
 
 public class Main {
 
 
     public static void main(String[] args) {
-        System.out.println(Task4.progression(5,3,4));
+        System.out.println(Task4.SumGeometricElements(100, 0.5, 20));
 
     }
